requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.2.0
Django==5.1.1
django-js-asset==2.2.0
django-tinymce==4.1.0
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
